+++ spec/fixtures/cookbooks/oneview_test/recipes/storage_pool_add_if_missing_with_name.rb
@@ -1,3 +1,7 @@
+#
+# Cookbook Name:: oneview_test
+# Recipe:: storage_pool_add_with_name
+#
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -8,29 +12,9 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+#
 
-my_client = {
-  url: '',
-  user: '',
-  password: ''
-}
-
-# Example: Adds storage pool if it is not added in oneview using
-# the storage system name
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  storage_system 'ThreePAR7200-8147'
-end
-
-# Example: Adds storage pool if it is not added in oneview using
-# the storage system hostname
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  storage_system '172.XX.XX.XX'
-end
-
-# Example: Removes storage pool from oneview
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  action :remove
+oneview_storage_pool 'StoragePool2' do
+  client node['oneview_test']['client']
+  storage_system 'StorageSystem1'
 end